fix(protocol): carry NamedColor discriminant as u16 so default cells aren't red
The wire grid encoded a cell's named colour as `WireColor::Named(u8)`, but
alacritty's specials are Foreground=256 and Background=257. The daemon's
`n as u8` truncated them to 0 and 1, so the client decoded every default
background as ansi16[1] (ANSI red) and default text as ansi16[0] (black) —
every daemon-rendered terminal rendered black-on-red.

Widen `WireColor::Named` to u16 so the discriminant rides the wire intact,
encode it verbatim in the daemon framer, and decode against alacritty's real
discriminants on the client (NamedColor::Foreground/Background as u16) instead
of the hard-coded 16/17 that never matched. Regression tests on both halves.

The u16 change re-bumps the build-generated SCHEMA_HASH, so a stale daemon is
refused by the handshake rather than silently mis-rendering.

+++ crates/micold-client/src/grid.rs
@@ -269,7 +269,7 @@ mod tests {
         }
     }
 
-    fn style(fg: u8) -> WireStyle {
+    fn style(fg: u16) -> WireStyle {
         WireStyle {
             fg: WireColor::Named(fg),
             bg: WireColor::Named(0),
